refactor : 이미지 파일 export하는 파일 index 만들고, import 경로 수정
