+++ src/assets/index.ts
@@ -1,0 +1,6 @@
+export { default as BackIcon } from "./BackIcon.png";
+export { default as CheckedButtonIcon } from "./CheckedButtonIcon.png";
+export { default as InfoIcon } from "./InfoIcon.png";
+export { default as MinusButtonIcon } from "./MinusButtonIcon.png";
+export { default as PlusButtonIcon } from "./PlusButtonIcon.png";
+export { default as UncheckedButtonIcon } from "./UncheckedButtonIcon.png";
